chore: update circular deps check depth to 5 WD-34344
