+++ src/components/forms/FormSubmitBtn.tsx
@@ -1,7 +1,7 @@
 import type { FC } from "react";
 import { ActionButton, useListener } from "@canonical/react-components";
 import { pluralize } from "util/instanceBulkActions";
-import type { ConfigurationRowFormikProps } from "components/ConfigurationRow";
+import type { ConfigurationRowFormikProps } from "types/forms/configurationRow";
 import { getFormChangeCount } from "util/formChangeCount";
 import { unstable_usePrompt as usePrompt } from "react-router";
