+++ src/pages/instances/CreateInstance.tsx
@@ -80,7 +80,8 @@ import YamlNotification from "components/forms/YamlNotification";
 import ProxyDeviceForm from "components/forms/ProxyDeviceForm";
 import ResourceLabel from "components/ResourceLabel";
 import type { InstanceIconType } from "types/instance";
-import BootForm, { bootPayload } from "components/forms/BootForm";
+import BootForm from "components/forms/BootForm";
+import { bootPayload } from "util/instanceBoot";
 import { useProfiles } from "context/useProfiles";
 import { sshKeyPayload } from "components/forms/SshKeyForm";
 import usePanelParams, { panels } from "util/usePanelParams";
