+++ src/pages/instances/InstanceDetailPanelContent.tsx
@@ -13,7 +13,8 @@ import { useInstanceLoading } from "context/instanceLoading";
 import { useIsClustered } from "context/useIsClustered";
 import InstanceMACAddresses from "pages/instances/InstanceMACAddresses";
 import ResourceLink from "components/ResourceLink";
-import { getImageLink, getInstanceType } from "util/instances";
+import { getInstanceType } from "util/instances";
+import { getImageLink } from "util/instanceImage";
 import ProfileRichChip from "pages/profiles/ProfileRichChip";
 import DevicesSummaryList from "components/DevicesSummaryList";
 import type { LxdDevices } from "types/device";
