+++ src/pages/instances/actions/AttachIsoBtn.tsx
@@ -16,7 +16,7 @@ import CustomIsoModal from "pages/images/CustomIsoModal";
 import type { FormDiskDevice } from "types/formDevice";
 import { remoteImageToIsoDevice } from "util/formDevices";
 import { useEventQueue } from "context/eventQueue";
-import { instanceLinkFromOperation } from "util/instances";
+import { instanceLinkFromOperation } from "util/operations";
 import ResourceLink from "components/ResourceLink";
 import { useInstanceEntitlements } from "util/entitlements/instances";
 import { InstanceRichChip } from "../InstanceRichChip";
