+++ src/pages/profiles/CreateProfile.tsx
@@ -68,7 +68,8 @@ import YamlSwitch from "components/forms/YamlSwitch";
 import YamlNotification from "components/forms/YamlNotification";
 import ProxyDeviceForm from "components/forms/ProxyDeviceForm";
 import ProfileRichChip from "pages/profiles/ProfileRichChip";
-import BootForm, { bootPayload } from "components/forms/BootForm";
+import BootForm from "components/forms/BootForm";
+import { bootPayload } from "util/instanceBoot";
 import { sshKeyPayload } from "components/forms/SshKeyForm";
 import usePanelParams, { panels } from "util/usePanelParams";
 import NetworkDevicePanel from "components/forms/NetworkDevicesForm/edit/NetworkDevicePanel";
