+++ src/util/instanceBoot.tsx
@@ -1,0 +1,17 @@
+import type { BootFormValues } from "types/forms/instanceAndProfile";
+import { getInstanceField } from "util/instanceConfigFields";
+
+export const bootPayload = (values: BootFormValues) => {
+  return {
+    [getInstanceField("boot_autostart")]: values.boot_autostart?.toString(),
+    [getInstanceField("boot_autostart_delay")]:
+      values.boot_autostart_delay?.toString(),
+    [getInstanceField("boot_autostart_priority")]:
+      values.boot_autostart_priority?.toString(),
+    [getInstanceField("boot_host_shutdown_timeout")]:
+      values.boot_host_shutdown_timeout?.toString(),
+    [getInstanceField("boot_mode")]: values.boot_mode?.toString(),
+    [getInstanceField("boot_stop_priority")]:
+      values.boot_stop_priority?.toString(),
+  };
+};
