+++ src/util/instances.tsx
@@ -1,37 +1,13 @@
-import type { LxdOperationResponse } from "types/operation";
-import { getInstanceName } from "./operations";
-import type { ReactNode } from "react";
 import type { AbortControllerState } from "util/helpers";
 import { checkDuplicateName } from "util/helpers";
 import { ROOT_PATH } from "util/rootPath";
 import * as Yup from "yup";
 import type { LxdInstance } from "types/instance";
-import { useImagesInProject } from "context/useImages";
-import ResourceLabel from "components/ResourceLabel";
-import ResourceLink from "components/ResourceLink";
-import { InstanceRichChip } from "pages/instances/InstanceRichChip";
 import { instanceCreationTypes } from "./instanceOptions";
 import { getInstanceLocation } from "./instanceLocation";
 import type { InstanceAndProfileFormikProps } from "types/forms/instanceAndProfileFormProps";
 
 export const CLUSTER_GROUP_PREFIX = "@";
-
-export const instanceLinkFromOperation = (args: {
-  operation?: LxdOperationResponse;
-  project?: string;
-}): ReactNode | undefined => {
-  const { operation, project } = args;
-  const instanceName = getInstanceName(operation?.metadata);
-  if (!instanceName) {
-    return;
-  }
-  return (
-    <InstanceRichChip
-      instanceName={instanceName}
-      projectName={project || "default"}
-    />
-  );
-};
 
 export const linkForInstanceDetail = (name: string, project?: string) => {
   return `${ROOT_PATH}/ui/project/${encodeURIComponent(project ?? "default")}/instance/${encodeURIComponent(name)}`;
@@ -74,30 +50,6 @@ export const instanceNameValidation = (
 
 export const getInstanceKey = (instance: LxdInstance) => {
   return `${instance.name} ${instance.project}`;
-};
-
-export const getImageLink = (instance: LxdInstance) => {
-  const { data: images = [] } = useImagesInProject(instance.project);
-  const imageDescription = instance.config["image.description"];
-  const imageFound = images?.some(
-    (image) => image.properties?.description === imageDescription,
-  );
-
-  if (!imageDescription) {
-    return "-";
-  }
-
-  if (!imageFound) {
-    return <ResourceLabel type="image" value={imageDescription} />;
-  }
-
-  return (
-    <ResourceLink
-      type="image"
-      value={imageDescription}
-      to={`${ROOT_PATH}/ui/project/${encodeURIComponent(instance.project)}/images`}
-    />
-  );
 };
 
 export const getInstanceMacAddresses = (instance: LxdInstance) => {
